minor navbar improvements
