+++ client/src/components/Dropdown.js
@@ -1,0 +1,25 @@
+import { useState, useEffect, useRef } from 'react';
+
+const useDropdown = () => {
+    const [hoveredDropdown, setHoveredDropdown] = useState(null);
+    const hoverTimeoutRef = useRef(null);
+
+    const handleMouseEnterDropdown = (menuName) => () => {
+        clearTimeout(hoverTimeoutRef.current);
+        setHoveredDropdown(menuName);
+    };
+
+    const handleMouseLeaveDropdown = () => {
+        hoverTimeoutRef.current = setTimeout(() => {
+            setHoveredDropdown(null);
+        }, 75);
+    };
+
+    useEffect(() => {
+        return () => clearTimeout(hoverTimeoutRef.current);
+    }, []);
+
+    return { hoveredDropdown, handleMouseEnterDropdown, handleMouseLeaveDropdown };
+};
+
+export default useDropdown;
