+++ client/src/components/NavBar/NavBar.css
@@ -15,8 +15,7 @@
 
 .menu-item {
     position: relative; 
-    margin-right: 10px;
-    margin-left: 10px;
+    margin: 0px 10px;
 }
 
 .menu-item a, .navbar-right a {
@@ -100,4 +99,14 @@
     text-decoration: none;
     color: black;
     display: block;
+}
+
+@media (max-width: 768px) {
+    .navbar {
+        padding: 5px 20px;
+    }
+
+    .search-input {
+        width: 150px; /* Adjust input width on smaller screens */
+    }
 }
