+++ client/src/components/SportsNavBar/SportsNavBar.css
@@ -8,7 +8,7 @@
     height: 35px;
 }
 
-.nav-link, .leagues-btn {
+.nav-link {
     text-decoration: none;
     color: white;
     margin-right: 10px;
@@ -18,40 +18,55 @@
     transition: background-color 0.2s, color 0.2s;
 }
 
-.nav-link:hover, .nav-link:focus, .leagues-btn:hover, .leagues-btn:focus {
+.nav-link:hover, .nav-link:focus {
     background-color: #5A5A5A;
     color: #DDDDDD;
 }
 
-.leagues-dropdown {
-    position: relative;
-    display: inline-block;
-}
-
-.dropdown-content {
-    display: none;
+.dropdown-right {
     position: absolute;
-    right: 0;
-    background-color: #333;
+    top: 155%; /* Align the top of the dropdown with the bottom of the menu-item */
+    left: -170%;
+    background-color: #c5c5c5;
     min-width: 160px;
-    box-shadow: 0 8px 16px 0 rgba(0,0,0,0.2);
+    box-shadow: 0px 8px 16px 0px rgba(0,0,0,0.2);
     z-index: 1;
 }
 
-.leagues-dropdown:hover .dropdown-content {
-    display: block;
+.dropdown-right::before {
+    content: "";
+    position: absolute;
+    top: -12px; /* Position the arrow above the dropdown */
+    left: 125px; /* Adjust this value to position the arrow correctly under the "Sports" header */
+    border-left: 12px solid transparent; /* Creates the left side of the arrow */
+    border-right: 12px solid transparent; /* Creates the right side of the arrow */
+    border-bottom: 12px solid #c5c5c5; /* Creates the base of the arrow, should match the dropdown background color */
+    z-index: 2;
 }
 
-.dropdown-content a {
-    color: white;
+.dropdown-right ul {
+    list-style: none;
+    padding: 0;
+    margin: 0;
+}
+
+.dropdown-right ul li {
     padding: 12px 16px;
+    text-align: left;
+}
+
+.dropdown-right ul li:hover {
+    background-color: #ddd;
+}
+
+.dropdown-right ul li a {
     text-decoration: none;
+    color: black;
     display: block;
 }
 
-.dropdown-content a:hover {
-    background-color: #ddd;
-    color: black;
+@media (max-width: 768px) {
+    .sports-navbar {
+        padding: 5px 20px;
+    }
 }
-
-
